+++ staj-project/src/main/java/com/hipatiya/staj/repository/ProductRepository.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 public interface ProductRepository extends JpaRepository<Product,Long> {
     Optional<Product> findByName(String name);
 
-    List<Product> findByPriceGreaterThan(Integer price);
+    List<Product> findByPriceGreaterThan(double price);
 
     @Query("SELECT p.name AS name, p.price AS price FROM Product p")
     List<ProductView> findProductViews();
